Created separate file for parallelisation tools, further tested and made them more stable.
