+++ main.py
@@ -5,6 +5,7 @@ import multiprocessing
 import gym
 from time import sleep
 import agents
+import parallelisation_tools as pll
 import game
 import data
 import logging
@@ -13,45 +14,17 @@ from game import *
 from config import *
 from worker import *
 
-PARALLEL = False
+PARALLEL = True
 
+@pll.Coordinator( PARALLEL )
+def play(sess, dataset, agent, env, thread="worker_1", GAMES_PER_ITER=GAMES_PER_ITER):
+    '''
+    Plays games against itself, sequentially or in parallel.
 
-def update_target_graph(from_scope,to_scope):
-    from_vars = tf.get_collection(tf.GraphKeys.TRAINABLE_VARIABLES, from_scope)
-    to_vars = tf.get_collection(tf.GraphKeys.TRAINABLE_VARIABLES, to_scope)
+    * comments:
 
-    op_holder = []
-    for from_var,to_var in zip(from_vars,to_vars):
-        op_holder.append(to_var.assign(from_var))
-    return op_holder
-
-
-class Coordinator:
-    def __init__(self, condition):
-        # We only want to use the coordinator if the algorithm is run in parallel
-        self.condition = condition
-        self.coord = tf.train.Coordinator()
-
-    def __call__(self, f):
-        if self.condition:
-            def wrapper(*args):
-                sess = args[0]
-                worker_threads = []
-                for worker in workers:
-                    worker.sess = sess
-                    worker_work = lambda: worker.work(f)
-                    t = threading.Thread(target=(worker_work))
-                    t.start()
-                    sleep(0.5)
-                    worker_threads.append(t)
-                self.coord.join(worker_threads)
-            return wrapper
-        return f
-
-@Coordinator( PARALLEL )
-def play(sess, dataset, agent, env, GAMES_PER_ITER=GAMES_PER_ITER):
-    '''Plays games against itself, sequentially or in parallel.
-    IMPORTANT: if you're changing it, make sure the first argument is still sess!'''
+    if you're changing it, make sure the first argument is still sess!
+    '''
     for i_game in range(GAMES_PER_ITER):
 
         f, a, r, o_s, a_s = game.create_play_data(sess, agent, env, without_net=WITHOUT_NET, quick_play=QUICK_PLAY)
@@ -64,48 +37,32 @@ def play(sess, dataset, agent, env, GAMES_PER_ITER=GAMES_PER_ITER):
             losses += 1
         if i_game % 20 == 0:
             agent.playing_log(i_game, wins, losses)
-            print("Game {0}. Agent: {1}, Opponent: {2}".format(i_game, a_s, o_s))
+            print("{0}. Game {1}. Agent: {2}, Opponent: {3}".format(thread, i_game, a_s, o_s))
 
-class Worker():
-    def __init__(self, name, agent, gpu_idx, dataset, test_on_cpu=False):
-        self.name = "worker_" + str(name)
-        self.train_data = game.train_set()
-        self.gpu_idx = str(gpu_idx)
-        self.dataset = dataset
-        self.sess = None
-        self.env = env = gym.make('Pong-v0')
-        if test_on_cpu:
-            self.device = '/cpu:'
-        else:
-            self.device = '/gpu:'
+def wrapped_agent(name):
+    '''
+    Wrapper used to make it easier to modify agents using modules (e.g. the logging module),
+    without there being multiple other dependencies in main.py that have to be changed should
+    an additional wrapper be introduced on the agent
 
-        #Create the local copy of the network and the tensorflow op to copy global paramters to local network
-        with tf.device(self.device+self.gpu_idx):
-            with tf.name_scope(name) as scope:
-                self.local_agent = agent
-                self.update_local_ops = update_target_graph('main_agent',self.name)
+    * arguments:
 
-    def work(self, play_function):
-        print("Starting play with worker "+self.name+" on GPU "+self.gpu_idx+"...")
-        play_function(self.sess, self.dataset, self.local_agent, self.env)
+    name
+        a str used as a prefix (or scope) for the tensorflow variables
 
-asynchronous = True
+    * comments:
 
-agent = agents.BasicAgent('main_agent', HIDDEN_SIZE, LEARNING_RATE)
-agent = logging_agent.Logging_Agent(agent)
+    N.A.
+
+    '''
+    initial_agent = agents.BasicAgent(name, HIDDEN_SIZE, LEARNING_RATE)
+    final_agent = logging_agent.Logging_Agent(initial_agent)
+    return final_agent
+
+agent = wrapped_agent('main_agent')
 dataset = data.Dataset()
 if PARALLEL:
-    num_workers = multiprocessing.cpu_count() # Set workers to number of available CPU threads
-    workers = []
-    # Create worker classes
-    num_gpus = 4
-    gpu_idx = 0
-    test_on_cpu = True
-    for i in range(num_workers):
-        workers.append(Worker(str(i), agents.BasicAgent("worker_" + str(i), HIDDEN_SIZE, LEARNING_RATE), gpu_idx, dataset, test_on_cpu=test_on_cpu))
-        gpu_idx += 1
-        if gpu_idx > num_gpus-1:
-            gpu_idx = 0
+
 
 
 env = gym.make('Pong-v0')
@@ -134,7 +91,7 @@ def main_function():
 
                 f,a,r = dataset.sample(TRAIN_BATCH_SIZE)
                 loss = agent.train(sess,f,a,r)
-                for worker in workers:
+                for worker in WORKERS:
                     worker.update_local_ops
                 if epoch % 500 == 0:
                     agent.epoch_log(sess, epoch, loss)
